Change `Job` media dimension return types from `int` to `float` and update corresponding tests for accuracy.

+++ src/Domain/Job.php
@@ -81,12 +81,12 @@ readonly class Job
         return $this->jobLengthMm;
     }
 
-    public function getMediaWidthMm(): int
+    public function getMediaWidthMm(): float
     {
         return $this->mediaWidthMm;
     }
 
-    public function getMediaLengthMm(): int
+    public function getMediaLengthMm(): float
     {
         return $this->mediaLengthMm;
     }
